ingestion-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
fiona==1.9.5
shapely==2.0.2
python-multipart==0.0.6
aiofiles==23.2.1
qdrant-client==1.7.0
